tests: cpu_load: move to tests/lib and cover both backends
The CPU load module now lives in lib/os/cpu_load, so move its test out of
tests/subsys/debug into tests/lib/cpu_load and rename the scenarios
accordingly.

Adapt the test to the unified module:

- Cover both backends. The idle-hook scenarios run on every architecture
  providing the hooks (Arm Cortex-M, Arm64, RISC-V); the
  runtime-statistics scenarios additionally run where the idle hooks are
  not available (x86, native_sim).
- Add SMP scenarios for both backends. The busy/idle sequence now runs in
  a worker thread pinned to a known CPU: a running thread cannot be
  pinned, and an unpinned worker may migrate, which would make the
  busy/idle attribution to a single CPU meaningless.
- Exercise the periodic report and the threshold callback on top of the
  runtime-statistics backend as well. Before the two modules merged, both
  features were only available with the idle-hook one.
- Add coverage for the new per-CPU getter cpu_load_get_cpu(), including
  its -EINVAL rejection of an out-of-range CPU id, for the current-CPU
  getter cpu_load_get() and for CPU_LOAD_PERMILLE_TO_PERCENT().
- Report the measured load in the assertion messages, so a failure says
  what was actually observed.
