+++ tests/lib/cpu_load/CMakeLists.txt
@@ -3,6 +3,6 @@
 cmake_minimum_required(VERSION 3.20.0)
 
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
-project(debug_cpu_load)
+project(cpu_load)
 
 target_sources(app PRIVATE src/main.c)
