+++ tests/lib/cpu_load/boards/nrf52840dk_nrf52840.overlay
