+++ tests/lib/cpu_load/boards/nrf54h20dk_nrf54h20_cpuapp.overlay
