+++ tests/lib/cpu_load/prj.conf
